requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
cmdstanpy==1.2.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
holidays==0.48
idna==3.7
importlib_resources==6.4.0
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
prophet==1.1.5
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
stanio==0.5.0
tenacity==8.3.0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
